move time formatting function to repo/mod.rs
